terminado

+++ WEB-INF/src/sekai/DeleteUser.java
@@ -21,48 +21,56 @@ public class DeleteUser extends HttpServlet {
 	@Override
 	protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException 
 	{
-		doPost(req, resp);
+		if (!db.databaseExists())
+			resp.sendRedirect(req.getContextPath() + "/installer");
+		else
+			doPost(req, resp);
 	}
 	
 	@Override
 	protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException 
 	{
-		PrintWriter out;
-        ST html;
-        HttpSession session;
+		if (!db.databaseExists())
+			resp.sendRedirect(req.getContextPath() + "/installer");
+		else
+		{
+			PrintWriter out;
+			ST html;
+			HttpSession session;
         
-        out = resp.getWriter();
-        session = req.getSession(false);
-        resp.setCharacterEncoding("UTF-8");
-        resp.setHeader("refresh","2;/sekai-project-gonzalo-racero-galan/panel");
-        resp.setHeader("Cache-Control","max-age=60");
-        html = new ST(HTMLTemplate.getDeleteUserMessage(),'$','$');
+			out = resp.getWriter();
+			session = req.getSession(false);
+			resp.setCharacterEncoding("UTF-8");
+			resp.setHeader("refresh","2;/sekai-project-gonzalo-racero-galan/panel");
+			resp.setHeader("Cache-Control","max-age=60");
+			html = new ST(HTMLTemplate.getDeleteUserMessage(),'$','$');
 
-        if (session != null)
-        {
-        	String usuario = (String) session.getAttribute("usuario_de_la_sesion");
+			if (session != null)
+			{
+				String usuario = (String) session.getAttribute("usuario_de_la_sesion");
         	
-        	if (usuario.equals("admin"))
-        	{
-        		String idUserDelete;
-        		Optional<Usuario> userDelete;
+				if (usuario.equals("admin"))
+				{
+					String idUserDelete;
+					Optional<Usuario> userDelete;
         		
-        		idUserDelete = req.getParameter("user_id");
-        		userDelete = db.findUsuarioById(idUserDelete);
+					idUserDelete = req.getParameter("user_id");
+					userDelete = db.findUsuarioById(idUserDelete);
         		
-        		if (userDelete.isPresent() && db.deleteUser(userDelete.get()) > 0)
-        			html.add("deleted_user", "El usuario " + idUserDelete + " fué eliminado con éxito");
-        		else
-        			html.add("deleted_user", "Se produjo un error al eliminar el usuario, es posible que este no exista");
+					if (userDelete.isPresent() && db.deleteUser(userDelete.get()) > 0)
+						html.add("deleted_user", "El usuario " + idUserDelete + " fué eliminado con éxito");
+					else
+						html.add("deleted_user", "Se produjo un error al eliminar el usuario, es posible que este no exista");
         		
-        		out.print(html.render());
-        	}
-        	else
-        		resp.sendRedirect(req.getContextPath() + "/login");
+					out.print(html.render());
+				}
+				else
+					resp.sendRedirect(req.getContextPath() + "/login");
         	
-        }
-        else
-        	resp.sendRedirect(req.getContextPath() + "/login");
+			}
+			else
+				resp.sendRedirect(req.getContextPath() + "/login");
         
+		}
 	}
 }
